TWILIO CHATBOT/requirements.txt
# LLM Twilio Callbot Dependencies

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Twilio Integration
twilio==8.10.0

# OpenAI Integration
openai==1.3.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Date/Time Handling
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.1.0

# Security
cryptography==41.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development Tools
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Production
gunicorn==21.2.0
gevent==23.7.0

# Optional: Database (for future use)
# SQLAlchemy==2.0.21
# psycopg2-binary==2.9.7

# Optional: Monitoring (for future use)
# prometheus-client==0.17.1
# statsd==4.0.1 